+++ packages/charts/src/chart_types/xy_chart/rendering/points.ts
@@ -19,6 +19,7 @@ import {
 import { Color } from '../../../common/colors';
 import { ScaleBand, ScaleContinuous } from '../../../scales';
 import { isFiniteNumber, isNil } from '../../../utils/common';
+import { inplaceInsertInSortedArray, SortedArray } from '../../../utils/data/data_processing';
 import { Dimensions } from '../../../utils/dimensions';
 import { BandedAccessorType, PointGeometry } from '../../../utils/geometry';
 import { PointStyle } from '../../../utils/themes/theme';
@@ -33,6 +34,11 @@ import {
 import { PointStyleAccessor, StackMode } from '../utils/specs';
 
 /** @internal */
+export function isolatedPointRadius(lineStrokeWidth: number) {
+  return lineStrokeWidth + 0.5;
+}
+
+/** @internal */
 export function renderPoints(
   shift: number,
   dataSeries: DataSeries,
@@ -41,7 +47,8 @@ export function renderPoints(
   panel: Dimensions,
   color: Color,
   pointStyle: PointStyle,
-  isolatedPointThemeStyle: PointStyle,
+  isolatedPointThemeStyle: Omit<PointStyle, 'radius'>,
+  lineStrokeWidth: number,
   isBandedSpec: boolean,
   markSizeOptions: MarkSizeOptions,
   useSpatialIndex: boolean,
@@ -60,7 +67,14 @@ export function renderPoints(
   const y1Fn = getY1ScaledValueFn(yScale);
   const y0Fn = getY0ScaledValueFn(yScale);
   const yDefined = isYValueDefinedFn(yScale, xScale);
-  const pointGeometries = dataSeries.data.reduce<PointGeometry[]>((acc, datum, dataIndex) => {
+
+  const needSorting = !markSizeOptions.enabled;
+
+  let style = buildPointGeometryStyles(color, pointStyle);
+  let isolatedPointStyle = buildPointGeometryStyles(color, isolatedPointThemeStyle);
+  let styleOverrides: Partial<PointStyle> | undefined = undefined;
+
+  const pointGeometries = dataSeries.data.reduce<SortedArray<PointGeometry>>((acc, datum, dataIndex) => {
     const { x: xValue, mark } = datum;
     const prev = dataSeries.data[dataIndex - 1];
     const next = dataSeries.data[dataIndex + 1];
@@ -73,19 +87,21 @@ export function renderPoints(
     if (Number.isNaN(x)) return acc;
 
     const yDatumKeyNames: Array<keyof Omit<FilledValues, 'x'>> = isBandedSpec ? ['y0', 'y1'] : ['y1'];
-
+    const seriesIdentifier: XYChartSeriesIdentifier = getSeriesIdentifierFromDataSeries(dataSeries);
+    const isPointIsolated = allowIsolated && isIsolatedPoint(dataIndex, dataSeries.data.length, yDefined, prev, next);
+    if (styleAccessor) {
+      styleOverrides = getPointStyleOverrides(datum, seriesIdentifier, isPointIsolated, styleAccessor);
+      style = buildPointGeometryStyles(color, pointStyle, styleOverrides);
+      isolatedPointStyle = buildPointGeometryStyles(color, isolatedPointThemeStyle, styleOverrides);
+    }
     yDatumKeyNames.forEach((yDatumKeyName, keyIndex) => {
       const valueAccessor = getYDatumValueFn(yDatumKeyName);
       const y = yDatumKeyName === 'y1' ? y1Fn(datum) : y0Fn(datum);
       const originalY = getDatumYValue(datum, keyIndex === 0, isBandedSpec, dataSeries.stackMode);
-      const seriesIdentifier: XYChartSeriesIdentifier = getSeriesIdentifierFromDataSeries(dataSeries);
-      const isPointIsolated = allowIsolated && isIsolatedPoint(dataIndex, dataSeries.data.length, yDefined, prev, next);
-      const styleOverrides = getPointStyleOverrides(datum, seriesIdentifier, isPointIsolated, styleAccessor);
-      const style = buildPointGeometryStyles(color, pointStyle, styleOverrides);
-      const isolatedPointStyle = buildPointGeometryStyles(color, isolatedPointThemeStyle, styleOverrides);
+
       // if radius is defined with the mark, limit the minimum radius to the theme radius value
       const radius = isPointIsolated
-        ? isolatedPointThemeStyle.radius
+        ? isolatedPointRadius(lineStrokeWidth)
         : markSizeOptions.enabled
           ? Math.max(getRadius(mark), pointStyle.radius)
           : styleOverrides?.radius ?? pointStyle.radius;
@@ -95,7 +111,7 @@ export function renderPoints(
         y: y === null ? NaN : y,
         radius,
         color,
-        style: isolatedPointThemeStyle.visible && isPointIsolated ? isolatedPointStyle : style,
+        style: isPointIsolated ? isolatedPointStyle : style,
         value: {
           x: xValue,
           y: originalY,
@@ -119,7 +135,11 @@ export function renderPoints(
         yScale.isValueInDomain(valueAccessor(datum)) &&
         !isDatumFilled(datum)
       ) {
-        acc.push(pointGeometry);
+        if (needSorting) {
+          inplaceInsertInSortedArray(acc, pointGeometry, (p) => p?.radius ?? NaN);
+        } else {
+          acc.push(pointGeometry);
+        }
       }
     });
     return acc;