+++ storybook/stories/mixed/6_fitting_stacked.story.tsx
@@ -135,7 +135,7 @@ export const Example: ChartsStory = (_, { title, description }) => {
         theme={{
           areaSeriesStyle: {
             point: {
-              visible: true,
+              visible: 'always',
             },
           },
         }}