+++ storybook/stories/mixed/7_marks.story.tsx
@@ -56,7 +56,12 @@ export const Example: ChartsStory = (_, { title, description }) => {
           markSizeRatio,
           areaSeriesStyle: {
             point: {
-              visible: true,
+              visible: 'always',
+            },
+          },
+          lineSeriesStyle: {
+            point: {
+              visible: 'always',
             },
           },
         }}